// packages/stable/sheets-sort/src/services/interface.ts
/**
 * Copyright 2023-present DreamNum Co., Ltd.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import type { IRange } from '@univerjs/core';

export interface ISortOption {
    range: IRange;
    orderRules: IOrderRule[];
    hasTitle?: boolean;
}

export enum SortType {
    DESC = 'desc', // Z-A
    ASC = 'asc', // A-Z
}

export interface IOrderRule {
    type: SortType;
    colIndex: number;
}